requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
yagmail==0.15.293
python-dotenv==1.0.1
psycopg2-binary==2.9.9
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
bcrypt==4.1.2
